+++ cogment/context.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from posixpath import join
 import grpc
 import grpc.aio  # type: ignore
 from prometheus_client import start_http_server as start_prometheus_server
@@ -24,109 +23,57 @@ import cogment.api.environment_pb2_grpc as env_grpc_api
 import cogment.api.hooks_pb2_grpc as hooks_grpc_api
 import cogment.api.datalog_pb2_grpc as datalog_grpc_api
 import cogment.api.trial_datastore_pb2_grpc as datastore_grpc_api
+import cogment.api.directory_pb2_grpc as directory_grpc_api
 
+import cogment.endpoints as ep
+from cogment.directory import Directory, ServiceType
 from cogment.actor import ActorSession
 from cogment.environment import EnvironmentSession
 from cogment.prehook import PrehookSession
 from cogment.datalog import DatalogSession
 from cogment.datastore import Datastore
 from cogment.control import Controller
-from cogment.errors import CogmentError
 from cogment.agent_service import AgentServicer, get_actor_impl
 from cogment.client_service import ClientServicer
 from cogment.env_service import EnvironmentServicer
 from cogment.hooks_service import PrehookServicer
 from cogment.datalog_service import DatalogServicer
+from cogment.errors import CogmentError
 from cogment.utils import logger
+from cogment.version import __version__
 
-from typing import Callable, Awaitable, Dict, List, Any
+from typing import Callable, Awaitable, Dict, List, Any, Tuple
 from types import ModuleType
-import os
 import asyncio
 from types import SimpleNamespace
+import socket
+import urllib.parse as urlpar
 
 
 DEFAULT_PROMETHEUS_PORT = 8000
+_ADDITIONAL_REGISTRATION_ITEMS = {"__registration_source" : "PythonSDK-Implicit", "__version" : __version__}
 
 
-class Endpoint:
-    """Class representing a remote Cogment endpoint where to connect."""
-
-    def __init__(self, url: str):
-        self.url = url
-        self.private_key = None
-        self.root_certificates = None
-        self.certificate_chain = None
-
-    def __str__(self):
-        result = f"Endpoint: url = {self.url}, private_key = {self.private_key}"
-        result += f", root_certificates = {self.root_certificates}, certificate_chain = {self.certificate_chain}"
-        return result
-
-    def set_from_files(self, private_key_file=None, root_certificates_file=None, certificate_chain_file=None):
-        try:
-            if private_key_file:
-                with open(private_key_file) as fl:
-                    self.private_key = fl.read()
-                if not self.private_key:
-                    self.private_key = None
-
-            if root_certificates_file:
-                with open(root_certificates_file) as fl:
-                    self.root_certificates = fl.read()
-                if not self.root_certificates:
-                    self.root_certificates = None
-
-            if certificate_chain_file:
-                with open(certificate_chain_file) as fl:
-                    self.certificate_chain = fl.read()
-                if not self.certificate_chain:
-                    self.certificate_chain = None
-
-        except Exception:
-            logger.error(f"Failed loading file from CWD: {os.getcwd()}")
-            raise
-
-
-class ServedEndpoint:
-    """Class representing a local Cogment endpoint where others can connect."""
-
-    def __init__(self, port: int):
-        self.port = port
-        self.private_key_certificate_chain_pairs = None
-        self.root_certificates = None
-
-    # def set_from_files(private_key_certificate_chain_pairs_file=None, root_certificates_file=None):
-    # TODO: This function would need to parse the PEM encoded `private_key_certificate_chain_pairs_file`
-    #       to create the list of tuples required (see simpler version in `Endpoint` class above).
-
-    def __str__(self):
-        result = f"ServedEndpoint: port = {self.port}"
-        result += f", private_key_certificate_chain_pairs = {self.private_key_certificate_chain_pairs}"
-        result += f", root_certificates = {self.root_certificates}"
-        return result
-
-
-def _make_client_channel(endpoint: Endpoint):
-    if endpoint.url[:7] == "grpc://":
-        url = endpoint.url[7:]
+def _make_client_channel(grpc_endpoint: ep.Endpoint):
+    parsed_url = urlpar.urlparse(grpc_endpoint.url)
+    if parsed_url.scheme == ep.GRPC_SCHEME:
+        url = parsed_url.netloc
     else:
-        logger.warning(f"Endpoint URL must be of gRPC type (start with 'grpc://') [{endpoint.url}]")
-        url = endpoint.url
+        raise CogmentError(f"Invalid endpoint scheme (must be 'grpc') [{grpc_endpoint.url}]")
 
-    if endpoint.private_key is None:
+    if not grpc_endpoint.using_ssl():
         channel = grpc.aio.insecure_channel(url)
     else:
-        if endpoint.root_certificates:
-            root = bytes(endpoint.root_certificates, "utf-8")
+        if grpc_endpoint.root_certificates:
+            root = bytes(grpc_endpoint.root_certificates, "utf-8")
         else:
             root = None
-        if endpoint.private_key:
-            key = bytes(endpoint.private_key, "utf-8")
+        if grpc_endpoint.private_key:
+            key = bytes(grpc_endpoint.private_key, "utf-8")
         else:
             key = None
-        if endpoint.certificate_chain:
-            certs = bytes(endpoint.certificate_chain, "utf-8")
+        if grpc_endpoint.certificate_chain:
+            certs = bytes(grpc_endpoint.certificate_chain, "utf-8")
         else:
             certs = None
         creds = grpc.ssl_channel_credentials(root, key, certs)
@@ -135,21 +82,21 @@ def _make_client_channel(endpoint: Endpoint):
     return channel
 
 
-def _make_server(endpoint: ServedEndpoint):
+def _make_server(served_endpoint: ep.ServedEndpoint):
     server = grpc.aio.server()
 
-    address = f"[::]:{endpoint.port}"
-    if endpoint.private_key_certificate_chain_pairs is None:
+    address = f"[::]:{served_endpoint.port}"
+    if not served_endpoint.using_ssl():
         server.add_insecure_port(address)
     else:
-        if endpoint.root_certificates:
+        if served_endpoint.root_certificates:
             require_client_auth = True
-            root = bytes(endpoint.root_certificates, "utf-8")
+            root = bytes(served_endpoint.root_certificates, "utf-8")
         else:
             require_client_auth = False
             root = None
         certs = []
-        for (key, chain) in endpoint.private_key_certificate_chain_pairs:
+        for (key, chain) in served_endpoint.private_key_certificate_chain_pairs:
             certs.append((bytes(key, "utf-8"), bytes(chain, "utf-8")))
         if not certs:
             certs = None
@@ -163,12 +110,14 @@ def _make_server(endpoint: ServedEndpoint):
 class Context:
     """Top level class for the Cogment library from which to obtain all services."""
 
-    def __init__(self, user_id: str, cog_settings: ModuleType, asyncio_loop=None, prometheus_registry=REGISTRY):
+    def __init__(self, user_id: str, cog_settings: ModuleType, asyncio_loop=None,
+                       prometheus_registry=REGISTRY, directory_endpoint: ep.Endpoint = None,
+                       directory_auth_token: str = None):
         self._user_id = user_id
         self._actor_impls: Dict[str, SimpleNamespace] = {}
         self._env_impls: Dict[str, SimpleNamespace] = {}
-        self._prehook_impl: Callable[[PrehookSession], Awaitable[None]] = None
-        self._datalog_impl: Callable[[DatalogSession], Awaitable[None]] = None
+        self._prehook_impl: SimpleNamespace = None
+        self._datalog_impl: SimpleNamespace = None
         self._grpc_server = None  # type: Any
         self._prometheus_registry = prometheus_registry
         self._cog_settings = cog_settings
@@ -183,6 +132,17 @@ class Context:
         else:
             self.asyncio_loop = asyncio_loop
 
+        if directory_endpoint is not None:
+            try:
+                channel = _make_client_channel(directory_endpoint)
+            except CogmentError as exc:
+                raise CogmentError(f"Directory endpoint: {exc}")
+
+            stub = directory_grpc_api.DirectorySPStub(channel)
+            self._directory = Directory(stub, directory_auth_token)
+        else:
+            self._directory = None
+
     def __str__(self):
         result = f"Cogment Context: user id = {self._user_id}"
         return result
@@ -190,44 +150,132 @@ class Context:
     def register_actor(self,
                        impl: Callable[[ActorSession], Awaitable[None]],
                        impl_name: str,
-                       actor_classes: List[str] = []):
+                       actor_classes: List[str] = [], properties: Dict[str, str] = {}):
+
         if self._grpc_server is not None:
             # We could accept "client" actor registration after the server is started, but it is not worth it
             raise CogmentError("Cannot register an actor after the server is started")
         if impl_name in self._actor_impls:
             raise CogmentError(f"The actor implementation name must be unique: [{impl_name}]")
+        if ep.ACTOR_CLASS_PROPERTY_NAME in properties:
+            raise CogmentError(f"Actor property [{ep.ACTOR_CLASS_PROPERTY_NAME}] is reserved for internal use")
+        if ep.IMPLEMENTATION_PROPERTY_NAME in properties:
+            raise CogmentError(f"Actor property [{ep.IMPLEMENTATION_PROPERTY_NAME}] is reserved for internal use")
 
-        self._actor_impls[impl_name] = SimpleNamespace(impl=impl, actor_classes=actor_classes)
+        directory_properties = {}
+        directory_properties.update(properties)
+        directory_properties[ep.IMPLEMENTATION_PROPERTY_NAME] = impl_name
+        directory_properties.update(_ADDITIONAL_REGISTRATION_ITEMS)
+
+        if len(actor_classes) > 0:
+            directory_actor_classes = [ac for ac in actor_classes]
+        else:
+            logger.deprecated(f"The use of an empty list of actor_classes is deprecated")  # type: ignore
+            directory_actor_classes = []
+            for ac in self._cog_settings.actor_classes:
+                actor_classes.append(ac.name)
+            logger.warning(f"Implementation [{impl_name}] will be registered to the directory"
+                            f" for all actor classes [{actor_classes}]")
+
+        self._actor_impls[impl_name] = SimpleNamespace(
+            impl=impl, actor_classes=directory_actor_classes, properties=directory_properties)
 
     def register_environment(self,
                              impl: Callable[[EnvironmentSession], Awaitable[None]],
-                             impl_name: str = "default"):
+                             impl_name: str = "default", properties: Dict[str, str] = {}):
         if self._grpc_server is not None:
             raise CogmentError("Cannot register an environment after the server is started")
         if impl_name in self._env_impls:
             raise CogmentError(f"The environment implementation name must be unique: [{impl_name}]")
+        if ep.IMPLEMENTATION_PROPERTY_NAME in properties:
+            raise CogmentError(f"Environment property [{ep.IMPLEMENTATION_PROPERTY_NAME}] is reserved for internal use")
 
-        self._env_impls[impl_name] = SimpleNamespace(impl=impl)
+        directory_properties = {}
+        directory_properties.update(properties)
+        directory_properties[ep.IMPLEMENTATION_PROPERTY_NAME] = impl_name
+        directory_properties.update(_ADDITIONAL_REGISTRATION_ITEMS)
+
+        self._env_impls[impl_name] = SimpleNamespace(impl=impl, properties=directory_properties)
 
     def register_pre_trial_hook(self,
-                                impl: Callable[[PrehookSession], Awaitable[None]]):
+                                impl: Callable[[PrehookSession], Awaitable[None]],
+                                properties: Dict[str, str] = {}):
         if self._grpc_server is not None:
             raise CogmentError("Cannot register a pre-trial hook after the server is started")
         if self._prehook_impl is not None:
             raise CogmentError("Only one pre-trial hook service can be registered")
 
-        self._prehook_impl = impl
+        directory_properties = {}
+        directory_properties.update(properties)
+        directory_properties.update(_ADDITIONAL_REGISTRATION_ITEMS)
+        self._prehook_impl = SimpleNamespace(impl=impl, properties=directory_properties)
 
     def register_datalog(self,
-                         impl: Callable[[DatalogSession], Awaitable[None]]):
+                         impl: Callable[[DatalogSession], Awaitable[None]],
+                         properties: Dict[str, str] = {}):
         if self._grpc_server is not None:
             raise CogmentError("Cannot register a datalog after the server is started")
         if self._datalog_impl is not None:
             raise CogmentError("Only one datalog service can be registered")
 
-        self._datalog_impl = impl
+        directory_properties = {}
+        directory_properties.update(properties)
+        directory_properties.update(_ADDITIONAL_REGISTRATION_ITEMS)
+        self._datalog_impl = SimpleNamespace(impl=impl, properties=directory_properties)
 
-    async def serve_all_registered(self, served_endpoint: ServedEndpoint, prometheus_port=DEFAULT_PROMETHEUS_PORT):
+    async def _directory_deregistration(self, registered: List[Tuple[int, str]]):
+        if self._directory is None:
+            return
+
+        for item in registered:
+            try:
+                await self._directory.deregister_service(*item)
+            except Exception as exc:
+                logger.debug(f"Failed to deregister service id [{item[0]}] from directory: [{exc}]")
+
+        registered.clear()
+
+    async def _directory_registration(self, port, ssl):
+        registered: List[Tuple[int, str]] = []
+
+        if self._directory is None:
+            logger.debug(f"No directory for implicit service registration")
+            return registered
+
+        socket_addr = socket.gethostbyname(socket.gethostname())
+        logger.debug(f"Registering services available here [{socket_addr}] to Directory")
+
+        try:
+            for actor in self._actor_impls.values():
+                properties = actor.properties
+                for actor_class in actor.actor_classes:
+                    properties[ep.ACTOR_CLASS_PROPERTY_NAME] = actor_class
+                    res = await self._directory.register_host(
+                        ServiceType.ACTOR, socket_addr, port, ssl, properties)
+                    registered.append(res)
+
+            for env in self._env_impls.values():
+                res = await self._directory.register_host(
+                    ServiceType.ENVIRONMENT, socket_addr, port, ssl, env.properties)
+                registered.append(res)
+
+            if self._prehook_impl is not None:
+                res = await self._directory.register_host(
+                    ServiceType.HOOK, socket_addr, port, ssl, self._prehook_impl.properties)
+                registered.append(res)
+
+            if self._datalog_impl is not None:
+                res = await self._directory.register_host(
+                    ServiceType.DATALOG, socket_addr, port, ssl, self._datalog_impl.properties)
+                registered.append(res)
+
+        except Exception:
+            await self._directory_deregistration(registered)
+            raise
+
+        return registered
+
+    async def serve_all_registered(self, served_endpoint: ep.ServedEndpoint, prometheus_port=DEFAULT_PROMETHEUS_PORT):
         if (len(self._actor_impls) == 0 and len(self._env_impls) == 0 and
                 self._prehook_impl is None and self._datalog_impl is None):
             raise CogmentError("Nothing registered to serve!")
@@ -245,11 +293,11 @@ class Context:
             env_grpc_api.add_EnvironmentSPServicer_to_server(servicer, self._grpc_server)
 
         if self._prehook_impl is not None:
-            servicer = PrehookServicer(self._prehook_impl, self._cog_settings, self._prometheus_registry)
+            servicer = PrehookServicer(self._prehook_impl.impl, self._cog_settings, self._prometheus_registry)
             hooks_grpc_api.add_TrialHooksSPServicer_to_server(servicer, self._grpc_server)
 
         if self._datalog_impl is not None:
-            servicer = DatalogServicer(self._datalog_impl, self._cog_settings)
+            servicer = DatalogServicer(self._datalog_impl.impl, self._cog_settings)
             datalog_grpc_api.add_DatalogSPServicer_to_server(servicer, self._grpc_server)
 
         if self._prometheus_registry is not None and prometheus_port is not None:
@@ -257,25 +305,74 @@ class Context:
 
         await self._grpc_server.start()
         logger.debug(f"Context gRPC server at port [{served_endpoint.port}] for user [{self._user_id}] started")
-        await self._grpc_server.wait_for_termination()
-        logger.debug(f"Context gRPC server at port [{served_endpoint.port}] for user [{self._user_id}] exited")
 
-    def get_controller(self, endpoint: Endpoint):
+        directory_registered = await self._directory_registration(served_endpoint.port, served_endpoint.using_ssl())
+
+        try:
+            await self._grpc_server.wait_for_termination()
+            logger.debug(f"Context gRPC server at port [{served_endpoint.port}] for user [{self._user_id}] exited")
+
+        finally:
+            await self._directory_deregistration(directory_registered)
+
+    def _make_controller(self, endpoint):
         channel = _make_client_channel(endpoint)
         stub = orchestrator_grpc_api.TrialLifecycleSPStub(channel)
         return Controller(stub, self._user_id)
 
-    def get_datastore(self, endpoint: Endpoint):
+    async def _inquire_and_make_controller(self, endpoint):
+        inquired_endpoint = await self._directory.get_inquired_endpoint(endpoint, ServiceType.LIFE_CYCLE)
+        return self._make_controller(inquired_endpoint)
+
+    # TODO: The non-async part is only kept for backward compatibility,
+    #       remove it in a future (backward incompatible) release.
+    def get_controller(self, endpoint=ep.Endpoint()):
+        try:
+            parsed_url = urlpar.urlparse(endpoint.url)
+        except Exception as exc:
+            raise CogmentError(f"Endpoint [{endpoint.url}]: {exc}")
+
+        if parsed_url.scheme == ep.GRPC_SCHEME or self._directory is None:
+            return self._make_controller(endpoint)  # This returns a controller instance
+        else:
+            return self._inquire_and_make_controller(endpoint)  # This returns an awaitable object
+
+    def _make_datastore(self, endpoint):
         channel = _make_client_channel(endpoint)
         stub = datastore_grpc_api.TrialDatastoreSPStub(channel)
         return Datastore(stub, self._cog_settings)
 
-    async def join_trial(self, trial_id, endpoint: Endpoint, impl_name=None, actor_name=None, actor_class=None):
+    async def _inquire_and_make_datastore(self, endpoint):
+        inquired_endpoint = await self._directory.get_inquired_endpoint(endpoint, ServiceType.DATASTORE)
+        return self.get_datastore(inquired_endpoint)
+
+    # TODO: The non-async part is only kept for backward compatibility,
+    #       remove it in a future (backward incompatible) release.
+    def get_datastore(self, endpoint=ep.Endpoint()):
+        try:
+            parsed_url = urlpar.urlparse(endpoint.url)
+        except Exception as exc:
+            raise CogmentError(f"Endpoint [{endpoint.url}]: {exc}")
+
+        if parsed_url.scheme == ep.GRPC_SCHEME or self._directory is None:
+            return self._make_datastore(endpoint)  # This returns a datastore instance
+        else:
+            return self._inquire_and_make_datastore(endpoint)  # This returns an awaitable object
+
+    # Undocumented
+    # We may want to make it async to standardize with the future
+    # versions of 'get_controller' and 'get_datastore'
+    def get_directory(self, endpoint: ep.Endpoint, authentication_token: str = None):
+        channel = _make_client_channel(endpoint)
+        stub = directory_grpc_api.DirectorySPStub(channel)
+        return Directory(stub, authentication_token)
+
+    async def join_trial(self, trial_id, endpoint=ep.Endpoint(), impl_name=None, actor_name=None, actor_class=None):
         requested_class = None
         requested_name = None
         if impl_name is not None:
             # For backward compatibility
-            logger.warning(f"`join_trial` parameter `impl_name` is deprecated")
+            logger.deprecated(f"`join_trial` parameter `impl_name` is deprecated")
             if actor_name is None:
                 actor_impl = self._actor_impls[impl_name]
                 if len(actor_impl.actor_classes) == 0:
@@ -296,7 +393,12 @@ class Context:
         else:
             raise CogmentError(f"Actor name or actor class must be specified to join a trial")
 
-        channel = _make_client_channel(endpoint)
+        if self._directory is not None:
+            inquired_endpoint = await self._directory.get_inquired_endpoint(endpoint, ServiceType.CLIENT_ACTOR)
+        else:
+            inquired_endpoint = endpoint
+
+        channel = _make_client_channel(inquired_endpoint)
         stub = orchestrator_grpc_api.ClientActorSPStub(channel)
         servicer = ClientServicer(self._cog_settings, stub)