+++ cogment/utils.py
@@ -21,16 +21,10 @@ from cogment.version import __version__
 import logging
 import types
 
-logger = logging.getLogger("cogment.sdk")
-
 
 def _logger_trace(self, msg, *args, **kwargs):
     """Simplified trace log output method"""
     self.log(self.TRACE, msg, *args, **kwargs)
-
-
-setattr(logger, "TRACE", 5)
-setattr(logger, "trace", types.MethodType(_logger_trace, logger))
 
 
 def _logger_deprecated(self, msg):
@@ -40,6 +34,9 @@ def _logger_deprecated(self, msg):
         self.DEPRECATED_LOGGED.add(msg)
 
 
+logger = logging.getLogger("cogment.sdk")
+setattr(logger, "TRACE", 5)
+setattr(logger, "trace", types.MethodType(_logger_trace, logger))
 setattr(logger, "DEPRECATED_LOGGED", set())
 setattr(logger, "deprecated", types.MethodType(_logger_deprecated, logger))