+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Support for actor diconnection and response timeout parameters
+- Global Directory integration
+- Support for self registration to the Directory
 
 ## v2.2.2 - 2022-06-20